requirements.txt
azure-core
azure-identity==1.15.0
azure-search-documents==11.6.0b1
python-dotenv==1.0.1
pymongo==4.6.1
openai==1.10.0
pypdf==4.0.0
PyPDF2==3.0.1
tiktoken==0.5.2
openpyxl==3.1.2
xlrd==2.0.1
python-docx==1.1.0
Flask==2.2.2
langchain==0.1.8
langchain-openai==0.0.6
Flask-Session==0.5
Werkzeug==2.3.7
requests==2.32.3
identity==0.5.2
gunicorn
databricks-sql-connector==3.1.2
zeep==4.2.1
certifi==2024.7.4
python-certifi-win32==1.6.1
opensearch-py==2.6.0
pandas==2.0.0 
loguru
